Creating Plex module
