+++ GapsWeb/pom.xml
@@ -23,6 +23,12 @@
         </dependency>
 
         <dependency>
+            <groupId>com.jasonhhouse</groupId>
+            <artifactId>Plex</artifactId>
+            <version>0.3.1</version>
+        </dependency>
+
+        <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter</artifactId>
         </dependency>
