+++ Plex/src/main/java/com/jasonhhouse/plex/Part.java
@@ -1,4 +1,4 @@
-package com.jasonhhouse.gaps;
+package com.jasonhhouse.plex;
 
 import java.util.Objects;
 import javax.xml.bind.annotation.XmlAttribute;
